backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pyjwt==2.8.0
bcrypt==4.1.2
groq==0.13.0
youtube-transcript-api==1.2.4
requests==2.31.0
python-dotenv==1.0.1
gunicorn==22.0.0
